chat routes, create, delete, get

+++ petiBackend/storyControllers/deleteStoryController.js
@@ -1,12 +1,36 @@
-const deleteStoryController = async (req, res, next) => {
-    // get the id of story from url
-    const { storyId } = req.params;
+const User = require('../models/User');
+const Story = require('../models/Story');
+const { CustomError } = require('../middlewares/error');
+
+
+const deleteAStoryController = async (req, res, next) => {
     try {
+        // Get the userId from the verified token (in cookies)
+        const userId = req.userId;
+
+        // Throw error if no userId is found
+        if (!userId) {
+            throw new CustomError("You have to login first", 401);
+        }
+
+        // get story id from url
+        const storyId = req.params;
+
+        // check if storyId is available
+        if (!storyId) {
+            throw new CustomError("You forgot the storyId", 400);
+        }
+
         // check if story exist
-        const story = await User.findById(storyId);
+        const story = await Story.findById(storyId);
         // throw error if not found
         if (!story){
             throw new CustomError("Story not found", 404);
+        }
+        // check if user that created the story is 
+        // the same user that want to delete the story
+        if (story.user !== userId){
+            throw new CustomError("You did not create the story", 401);
         }
         // find and delete
         await Story.findByIdAndDelete(storyId);
@@ -20,9 +44,15 @@ const deleteStoryController = async (req, res, next) => {
 }
 
 const deleteUserStoriesController = async (req, res, next) => {
-    // get user id from url
-    const { userId } = req.params;
     try {
+        // Get the userId from the verified token (in cookies)
+        const userId = req.userId;
+
+        // Throw error if no userId is found
+        if (!userId) {
+            throw new CustomError("You have to login first", 401);
+        }
+
         // check if user exist
         const user = await User.findById(userId);
         // throw error if not found
@@ -37,4 +67,9 @@ const deleteUserStoriesController = async (req, res, next) => {
     } catch(error) {
         next(error);
     }
+}
+
+module.exports = {
+    deleteAStoryController,
+    deleteUserStoriesController
 }
